+++ click_extra/tabulate.py
@@ -19,8 +19,10 @@
 
 from __future__ import annotations
 
+import csv
 from functools import partial
 from gettext import gettext as _
+from io import StringIO
 
 import tabulate
 from click import Choice, echo
@@ -29,29 +31,85 @@ from cloup import option
 from .parameters import ExtraOption
 
 tabulate.MIN_PADDING = 0
+"""Neutralize spurious double-spacing in table rendering."""
+
+
+output_formats = sorted(
+    # Formats from tabulate.
+    list(tabulate._table_formats)
+    # Formats inherited from previous legacy cli-helpers dependency.
+    + ["csv", "vertical"]
+    # Formats derived from CSV dialects.
+    + [f"csv-{d}" for d in csv.list_dialects()]
+)
+"""All output formats supported by click-extra."""
+
+
+def get_csv_dialect(format_id: str) -> str | None:
+    """Extract, validate and normalize CSV dialect ID from format."""
+    assert format_id.startswith("csv")
+    # Defaults to excel rendering, like in Python's csv module.
+    dialect = "excel"
+    parts = format_id.split("-", 1)
+    assert parts[0] == "csv"
+    if len(parts) > 1:
+        dialect = parts[1]
+    return dialect
+
+
+def render_csv(tabular_data, headers=(), **kwargs):
+    with StringIO(newline="") as output:
+        writer = csv.writer(output, **kwargs)
+        writer.writerow(headers)
+        writer.writerows(tabular_data)
+        echo(output.getvalue(), nl=False, color=False)
+
+
+def render_vertical(tabular_data, headers=(), **kwargs):
+    """Re-implements cli-helpers' vertical table layout.
+
+    Source: https://github.com/dbcli/cli_helpers/blob/v2.3.0/cli_helpers/tabular_output/vertical_table_adapter.py
+    """
+    header_len = max(len(h) for h in headers)
+    padded_headers = [h.ljust(header_len) for h in headers]
+
+    for index, row in enumerate(tabular_data):
+        # 27 has been hardcoded in cli-helpers:
+        # https://github.com/dbcli/cli_helpers/blob/4e2c417f68bc07c72b508e107431569b0783c4ef/cli_helpers/tabular_output/vertical_table_adapter.py#L34
+        echo(f"{'*' * 27}[ {index + 1}. row ]{'*' * 27}")
+        for cell_label, cell_value in zip(padded_headers, row):
+            echo(f"{cell_label} | {cell_value}")
+
+
+def render_table(tabular_data, headers=(), **kwargs):
+    """Render a table with tabulate and output it via echo."""
+    defaults = {
+        "disable_numparse": True,
+        "numalign": None,
+    }
+    defaults.update(kwargs)
+    echo(tabulate.tabulate(tabular_data, headers, **defaults))
 
 
 class TableFormatOption(ExtraOption):
     """A pre-configured option that is adding a ``-t``/``--table-format`` flag to select
     the rendering style of a table."""
 
-    def print_table(self, *args, **kwargs):
-        """Print table via echo."""
-        defaults = {
-            "disable_numparse": True,
-            "numalign": None,
-        }
-        defaults.update(kwargs)
-        echo(tabulate.tabulate(*args, **defaults))
-
     def init_formatter(self, ctx, param, value):
-        """Attach a ready-to-use ``print_table()`` method to the context."""
-        ctx.print_table = partial(self.print_table, tablefmt=value)
+        """Attach a ready-to-use ``print_table(tabular_data, headers)`` method to the context."""
+        render_func = None
+        if value.startswith("csv"):
+            render_func = partial(render_csv, dialect=get_csv_dialect(value))
+        elif value == "vertical":
+            render_func = render_vertical
+        else:
+            render_func = partial(render_table, tablefmt=value)
+        ctx.print_table = render_func
 
     def __init__(
         self,
         param_decls=None,
-        type=Choice(sorted(tabulate._table_formats), case_sensitive=False),
+        type=Choice(output_formats, case_sensitive=False),
         default="rounded_outline",
         expose_value=False,
         help=_("Rendering style of tables."),